% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lkd_auth}
\alias{lkd_auth}
\title{Authorization in Linkedin API}
\usage{
lkd_auth(login = Sys.getenv("LKD_LOGIN"))
}
\arguments{
\item{login}{your Linkedin login}
}
\value{
No return value, just take API token
}
\description{
Authorization in Linkedin API. For more details see \href{https://learn.microsoft.com/en-us/linkedin/shared/authentication/authentication}{link}.
}
\examples{
\dontrun{
# set auth data
lkd_set_client_id('Your client id')
lkd_set_client_secret('Your client secret')
lkd_set_login('Your linkedin login')

lkd_auth()

}
}
