% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd_get_accounts_by_authenticated_user.R
\name{lkd_get_accounts_by_authenticated_user}
\alias{lkd_get_accounts_by_authenticated_user}
\title{Find Ad Accounts by Authenticated User
All ad accounts that an authenticated user has access}
\usage{
lkd_get_accounts_by_authenticated_user(start = 0, count = 1000)
}
\arguments{
\item{start}{Integer, paggination. The index of the first item you want results for.}

\item{count}{Integer, pagination. The number of items you want included on each page of results. There could be fewer items remaining than the value you specify.}
}
\value{
tibble with accounts list
}
\description{
Find Ad Accounts by Authenticated User
All ad accounts that an authenticated user has access
}
