% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd_get_campaigns.R
\name{lkd_get_campaigns}
\alias{lkd_get_campaigns}
\title{Get campaigns}
\usage{
lkd_get_campaigns(
  account_id = lkd_get_account_id(),
  test = FALSE,
  start = 0,
  count = 1000
)
}
\arguments{
\item{account_id}{your Linkedin Ad Account ID}

\item{test}{Searches for campaigns based on test or non-test status:
\itemize{
\item True: for test campaigns
\item False: for non-test campaigns If not specified, searches for both test and non-test campaigns.
}}

\item{start}{Integer, paggination. The index of the first item you want results for.}

\item{count}{Integer, pagination. The number of items you want included on each page of results. There could be fewer items remaining than the value you specify.}
}
\value{
tibble with campaign metadata
}
\description{
Get campaigns
}
