% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd_get_creatives.R
\name{lkd_get_creatives}
\alias{lkd_get_creatives}
\title{Get Creatives
You can search for creative content in order to get a collection of creatives matching}
\usage{
lkd_get_creatives(account_id = lkd_get_account_id(), start = 0, count = 1000)
}
\arguments{
\item{account_id}{your Linkedin Ad Account ID}

\item{start}{Integer, paggination. The index of the first item you want results for.}

\item{count}{Integer, pagination. The number of items you want included on each page of results. There could be fewer items remaining than the value you specify.}
}
\value{
tibble with creatives metadata
}
\description{
Get Creatives
You can search for creative content in order to get a collection of creatives matching
}
