% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lkd_remove_token}
\alias{lkd_remove_token}
\title{remove token}
\usage{
lkd_remove_token(login = lkd_get_login())
}
\arguments{
\item{login}{your login at Linkedin}
}
\value{
no return value, just remove token cacche
}
\description{
remove token
}
