% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lkd_set_account_id}
\alias{lkd_set_account_id}
\title{Set Ad Account ID for R session}
\usage{
lkd_set_account_id(account_id)
}
\arguments{
\item{account_id}{Id of your ad account}
}
\value{
no return value
}
\description{
Set Ad Account ID for R session
}
