% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lkd_set_api_version}
\alias{lkd_set_api_version}
\title{Set Linkedin API version
For dwtwils se next \href{https://learn.microsoft.com/en-us/linkedin/marketing/versioning?view=li-lms-2023-10}{link}}
\usage{
lkd_set_api_version(api_version)
}
\arguments{
\item{api_version}{Linkedin API version}
}
\value{
no return value
}
\description{
Set Linkedin API version
For dwtwils se next \href{https://learn.microsoft.com/en-us/linkedin/marketing/versioning?view=li-lms-2023-10}{link}
}
