% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lkd_set_client_id}
\alias{lkd_set_client_id}
\title{Set App Client ID
For more information see this \href{https://learn.microsoft.com/en-us/linkedin/marketing/quick-start?view=li-lms-2023-10#step-1-apply-for-api-access}{links}.}
\usage{
lkd_set_client_id(client_id)
}
\arguments{
\item{client_id}{you linkedin APP ID}
}
\value{
no return value
}
\description{
Set App Client ID
For more information see this \href{https://learn.microsoft.com/en-us/linkedin/marketing/quick-start?view=li-lms-2023-10#step-1-apply-for-api-access}{links}.
}
