% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lkd_set_client_secret}
\alias{lkd_set_client_secret}
\title{Set App Client secret
For more information see this \href{https://learn.microsoft.com/en-us/linkedin/marketing/quick-start?view=li-lms-2023-10#step-1-apply-for-api-access}{links}.}
\usage{
lkd_set_client_secret(client_secret)
}
\arguments{
\item{client_secret}{Your APP secret}
}
\value{
no return value
}
\description{
Set App Client secret
For more information see this \href{https://learn.microsoft.com/en-us/linkedin/marketing/quick-start?view=li-lms-2023-10#step-1-apply-for-api-access}{links}.
}
