% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lkd_set_login}
\alias{lkd_set_login}
\title{Set login for current R session}
\usage{
lkd_set_login(login)
}
\arguments{
\item{login}{Your Linkedin login}
}
\value{
no return value
}
\description{
Set login for current R session
}
