% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{lkd_set_token_path}
\alias{lkd_set_token_path}
\title{Set own token cache path}
\usage{
lkd_set_token_path(token_path)
}
\arguments{
\item{token_path}{Path to cache folder}
}
\value{
no return value
}
\description{
Set own token cache path
}
