% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd_get_campaign_groups.R
\name{lkd_get_campaign_groups}
\alias{lkd_get_campaign_groups}
\title{Get Ad campaign Groups
Campaign groups provide advertisers a way to manage status, budget, and performance across multiple related campaigns.}
\usage{
lkd_get_campaign_groups(account_id = lkd_get_account_id(), test = FALSE)
}
\arguments{
\item{account_id}{your Linkedin Ad Account ID}

\item{test}{Searches for campaigns based on test or non-test status:
\itemize{
\item True: for test campaigns
\item False: for non-test campaigns If not specified, searches for both test and non-test campaigns.
}}
}
\value{
tibble with campaign groups metadata
}
\description{
Get Ad campaign Groups
Campaign groups provide advertisers a way to manage status, budget, and performance across multiple related campaigns.
}
